docker/mqtt-client/requirements.txt
paho-mqtt
mysql-connector-python
paho-mqtt 
numpy 
matplotlib
roundrobin